feat: Extract CardView
