+++ Memorize/Memorize/EmojiMemoryGameView.swift
@@ -38,31 +38,6 @@ struct EmojiMemoryGameView: View {
     }
 }
 
-struct CardView: View {
-    let card: MemoryGame<String>.Card
-
-    init(_ card: MemoryGame<String>.Card) {
-        self.card = card
-    }
-
-    var body: some View {
-        ZStack {
-            let base = RoundedRectangle(cornerRadius: 12.0)
-            Group {
-                base.fill(.white)
-                base.strokeBorder(lineWidth: 2.0)
-                Text(card.content)
-                    .font(.system(size: 200))
-                    .minimumScaleFactor(0.01)
-                    .aspectRatio(1, contentMode: .fit)
-            }
-            .opacity(card.isFaceUp ? 1 : 0)
-            base.fill().opacity(card.isFaceUp ? 0 : 1)
-        }
-        .opacity(card.isFaceUp || !card.isMathced ? 1 : 0)
-    }
-}
-
 #Preview {
     EmojiMemoryGameView(viewModel: .init())
 }
